- Start adding support for manipulating and adding functions, which is a must to inject imports. - Added stub for encoding all common WebAssembly sections.
This commit holds the implementation for referecing functions, tables, types, which are not 100% bullet prof, it simply runs trough every of these collection and caches the index
of it within the obj._index property, this means that for example a function which has been removed will not warn where a call-site for it is encoded, if the _index property it may
result in a call-site to the function which are not at the given index, it may result in hard to track bugs in the resulting WebAssembly bytecode. The way to avoid this is to
lookup the index at encoding time, using the functions.indexOf(func) which results in longer encoding times but ensures that the output is valid.